test: split PDF visual caption scaffold coverage
